fix: show cross-project subtasks in List view (#1649)

+++ frontend/cypress/e2e/project/prepareProjects.ts
@@ -2,25 +2,27 @@ import {ProjectFactory} from '../../factories/project'
 import {TaskFactory} from '../../factories/task'
 import {ProjectViewFactory} from "../../factories/project_view";
 
-export function createDefaultViews(projectId) {
-	ProjectViewFactory.truncate()
+export function createDefaultViews(projectId: number, startViewId = 1, truncate: boolean = true) {
+	if (truncate) {
+		ProjectViewFactory.truncate()
+	}
 	const list = ProjectViewFactory.create(1, {
-		id: 1,
+		id: startViewId,
 		project_id: projectId,
 		view_kind: 0,
 	}, false)
 	const gantt = ProjectViewFactory.create(1, {
-		id: 2,
+		id: startViewId + 1,
 		project_id: projectId,
 		view_kind: 1,
 	}, false)
 	const table = ProjectViewFactory.create(1, {
-		id: 3,
+		id: startViewId + 2,
 		project_id: projectId,
 		view_kind: 2,
 	}, false)
 	const kanban = ProjectViewFactory.create(1, {
-		id: 4,
+		id: startViewId + 3,
 		project_id: projectId,
 		view_kind: 3,
 		bucket_configuration_mode: 1,
@@ -34,12 +36,19 @@ export function createDefaultViews(projectId) {
 	]
 }
 
-export function createProjects() {
-	const projects = ProjectFactory.create(1, {
-		title: 'First Project'
+export function createProjects(count: number = 1) {
+	const projects = ProjectFactory.create(count, {
+		title: i => count === 1 ? 'First Project' : `Project ${i + 1}`,
 	})
+
 	TaskFactory.truncate()
-	projects.views = createDefaultViews(projects[0].id)
+	ProjectViewFactory.truncate()
+
+	for (let i = 0; i < projects.length; i++) {
+		const views = createDefaultViews(projects[i].id, i * 4 + 1, false)
+		projects[i].views = views
+	}
+
 	return projects
 }
